% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcrawl.R
\name{crawl_description}
\alias{crawl_description}
\title{Crawl description from stocks}
\usage{
crawl_description(stocks)
}
\arguments{
\item{stocks}{a stocks tibble}
}
\value{
tibble description with
}
\description{
Crawl description from stocks
}
\examples{
crawl_stocks() \%>\% crawl_description()
}
