% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcrawl.R
\name{crawl_news}
\alias{crawl_news}
\title{Crawl news from stocks}
\usage{
crawl_news(stocks)
}
\arguments{
\item{n}{the number of stocks to crawl}
}
\value{
tibble with
}
\description{
Crawl news from stocks
}
\examples{
crawl_stocks() \%>\% crawl_description()
}
