% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcrawl.R
\name{crawl_stocks}
\alias{crawl_stocks}
\title{Crawl stocks up to the first n best}
\usage{
crawl_stocks(n = 20)
}
\arguments{
\item{n}{the number of stocks to crawl}
}
\description{
Crawl stocks up to the first n best
}
\examples{
crawl_stocks(10)
}
