% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcrawl.R
\name{get_description}
\alias{get_description}
\title{Gets description from a stock}
\usage{
get_description(ticker)
}
\arguments{
\item{n}{the number of stocks to crawl}
}
\description{
Gets description from a stock
}
\examples{
get_description('aapl')
}
