% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcrawl.R
\name{get_news}
\alias{get_news}
\title{Gets news from a stock}
\usage{
get_news(ticker)
}
\arguments{
\item{ticker}{stock symbol ticker}
}
\value{
character vector with headlines
}
\description{
Gets news from a stock
}
\examples{
get_news('aapl')
}
