% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senser_data.R
\name{senser_data}
\alias{senser_data}
\title{Retrieves a dataframe containing data from a Qlik Sense HyperCube}
\usage{
senser_data(host, port, app, hyperCubeDef, fields)
}
\arguments{
\item{host}{a URL where the q-risotto service runs, usually the Qlik Sense
Desktop (localhost) or Server IP address}

\item{port}{a port number where to connect q-risotto,}

\item{app}{a Qlik Sense application name to connect, usually a GUID for
applications on a Qlik Sense Server}

\item{hyperCubeDef}{a list of dimensions (field names) and measure
expressions (starting with '='), see q-risotto README for more details}

\item{fields}{a list of field names which will be assigned to the resulting
vectors}
}
\value{
a dataframe containing the resulting HyperCube data
}
\description{
This function makes a REST POST call to q-risotto (the Qlik Sense REST API
wrapper service) to retrieve data out of a Qlik Sense application. You will
find more details about q-risotto here:
https://github.com/ralfbecher/q-risotto
}
\details{
In the parameters, you define the Qlik Sense HyperCube by giving a set of
dimensions and measure expressions (Qlik syntax). The HyperCube will then be
created in the specified Qlik Sense application and the resulting data are
retrieved completely by paging calls.
}
\examples{
\dontrun{
host <- "http://localhost"
port <- 3000
app <- "Helpdesk Management.qvf"
hyperCubeDef <- c("Date.autoCalendar.Date", "=Count( {$<[Case Is Closed] ={'True'} >} Distinct \%CaseId )", "=Count( {$<[Status]={'New'} >} Distinct \%CaseId )")
fields <- c("Date", "Closed Cases", "New Cases")
senser_data(host=host,port=port,app=app,hyperCubeDef=hyperCubeDef,fields=fields)}
}
